wider

+++ src/n2n.py
@@ -801,8 +801,8 @@ class N2N(nn.Module):
                 # print(f'len: {len(listOfRunningMean)}')
                 # print(f'm1list: {m1list}')
 
-                nw1 = torch.cat(w1,dw1,dim=1)
-                nw2 = torch.cat(w2,dw2,dim=0)
+                nw1 = torch.cat(w1,torch.FloatTensor(dw1),dim=1)
+                nw2 = torch.cat(w2,torch.FloatTensor(dw2),dim=0)
                 nbn1rm = torch.cat(bn.running_mean, dbn1rm)
                 nbn1rv = torch.cat(bn.running_var, dbn1rv)
                 nbn1w = torch.cat(bn.weight.data, dbn1w)
